Refactor files to limit duplication
