+++ support-frontend/assets/components/subscriptionCheckouts/thankYou/returnSection.jsx
@@ -8,11 +8,15 @@ import { sendTrackingEventsOnClick } from 'helpers/subscriptions';
 
 import Content, { Divider, NarrowContent } from 'components/content/content';
 import AnchorButton from 'components/button/anchorButton';
+import type { SubscriptionProduct } from 'helpers/subscriptions';
 
+type PropTypes = {
+  subscriptionProduct: SubscriptionProduct,
+}
 
 // ----- Component ----- //
 
-function ReturnSection() {
+function ReturnSection(props: PropTypes) {
 
   return (
     <Content>
@@ -23,7 +27,7 @@ function ReturnSection() {
             appearance="secondary"
             aria-label="Return to The Guardian home page"
             href="https://theguardian.com"
-            onClick={sendTrackingEventsOnClick('checkout_return_home', 'Paper', null)}
+            onClick={sendTrackingEventsOnClick('checkout_return_home', props.subscriptionProduct, null)}
           >
             Return to The Guardian
           </AnchorButton>
