+++ support-frontend/assets/components/subscriptionCheckouts/thankYou/thankYou.module.scss
