+++ support-frontend/assets/pages/digital-subscription-checkout/thankYouPendingContent.jsx
