+++ support-frontend/assets/pages/paper-subscription-checkout/components/thankYou.jsx
@@ -18,8 +18,8 @@ import Text, { LargeParagraph, SansParagraph } from 'components/text/text';
 import { HeroWrapper } from 'components/productPage/productPageHero/productPageHero';
 import HeadingBlock from 'components/headingBlock/headingBlock';
 import { manageSubsUrl, myAccountUrl } from 'helpers/externalLinks';
-import typeof MarketingConsent from 'pages/paper-subscription-checkout/components/thankYou/marketingConsentContainer';
-import styles from 'pages/paper-subscription-checkout/components/thankYou/thankYou.module.scss';
+import typeof MarketingConsent from 'components/subscriptionCheckouts/thankYou/marketingConsentContainer';
+import styles from 'components/subscriptionCheckouts/thankYou/thankYou.module.scss';
 import { formatUserDate } from 'helpers/dateConversions';
 
 import { type FormFields, getFormFields } from 'helpers/subscriptionsForms/formFields';
@@ -56,7 +56,7 @@ const whatNext: {[FulfilmentOptions]: Element<*>} = {
         <OrderedList items={[
           <span>
             Look out for an email from us confirming your subscription.
-            It has everything you need to know about how manage it in the future.
+            It has everything you need to know about how to manage it in the future.
           </span>,
           <span>
             You will receive your personalised book of vouchers.
@@ -117,7 +117,7 @@ function ThankYouContent({
         </Text>
       </Content>
       <Content>
-        <Asyncronously loader={import('pages/paper-subscription-checkout/components/thankYou/marketingConsentContainer')}>
+        <Asyncronously loader={import('components/subscriptionCheckouts/thankYou/marketingConsentContainer')}>
           {(MktConsent: MarketingConsent) => (
 
             <MktConsent
