+++ support-frontend/assets/pages/paper-subscription-checkout/stage.jsx
@@ -9,10 +9,10 @@ import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
 
 import ProgressMessage from 'components/progressMessage/progressMessage';
 
+import { type Stage } from 'helpers/subscriptionsForms/formFields';
+import ThankYouContent from './components/thankYou';
 import CheckoutForm from './components/checkoutForm';
-import ReturnSection from 'pages/paper-subscription-checkout/components/thankYou/returnSection';
-import ThankYouContent from 'pages/paper-subscription-checkout/components/thankYou/thankYou';
-import type { Stage } from 'helpers/subscriptionsForms/formFields';
+import ReturnSection from 'components/subscriptionCheckouts/thankYou/returnSection';
 import type { WithDeliveryCheckoutState } from 'helpers/subscriptionsForms/subscriptionCheckoutReducer';
 
 // ----- Types ----- //
@@ -41,7 +41,7 @@ function CheckoutStage(props: PropTypes) {
       return (
         <div>
           <ThankYouContent isPending={false} />
-          <ReturnSection />
+          <ReturnSection subscriptionProduct="Paper" />
         </div>
       );
 
@@ -49,7 +49,7 @@ function CheckoutStage(props: PropTypes) {
       return (
         <div>
           <ThankYouContent isPending />
-          <ReturnSection />
+          <ReturnSection subscriptionProduct="Paper" />
         </div>
       );
