--- support-frontend/assets/pages/weekly-subscription-checkout/components/thankYou/thankYou.module.scss
@@ -1,40 +1,0 @@
-@import '~stylesheets/gu-sass/gu-sass';
-
-.hero {
-  background: gu-colour(brand-main);
-  border-top: 1px solid gu-colour(brand-pastel);
-  @include mq($from: tablet) {
-
-   :global(.component-heading-block) {
-      margin-top: 26em;
-    }
-
-    :global(.component-grid-picture) {
-      position: absolute;
-      top: -50px;
-      min-width: 100%;
-      height: 100%;
-      @include mq($from: wide) {
-        left: gu-span(2) * -1;
-      }
-    }
-    :global(.component-grid-picture__image) {
-      height: 100%;
-      display: block;
-    }
-  }
-
-  @include mq($until: tablet) {
-    :global(.component-grid-picture) {
-      max-height: 20em;
-      overflow: hidden;
-      display: block;
-    }
-    :global(.component-grid-picture__image) {
-      max-width: gu-span(8);
-      margin: auto;
-      float: none;
-      display: block;
-    }
-  }
-}
