+++ support-frontend/assets/pages/weekly-subscription-checkout/weeklySubscriptionCheckout.jsx
@@ -17,6 +17,7 @@ import SubscriptionTermsPrivacy
 import SubscriptionFaq from 'components/subscriptionFaq/subscriptionFaq';
 import 'stylesheets/skeleton/skeleton.scss';
 import CheckoutStage from './stage';
+import CheckoutForm from './components/weeklyCheckoutForm';
 import ConsentBanner from '../../components/consentBanner/consentBanner';
 import type { CommonState } from 'helpers/page/commonReducer';
 import { createWithDeliveryCheckoutReducer } from 'helpers/subscriptionsForms/subscriptionCheckoutReducer';
@@ -69,7 +70,9 @@ const content = (
         </Footer>
       }
     >
-      <CheckoutStage />
+      <CheckoutStage subscriptionProduct="GuardianWeekly">
+        <CheckoutForm />
+      </CheckoutStage>
       <ConsentBanner />
     </Page>
   </Provider>
